+++ packages/api/src/client/index.ts
@@ -1,39 +1,23 @@
+import 'setimmediate'
+import { EventEmitter } from 'events'
 import Queue from 'better-queue'
 // @ts-ignore
 import MemoryStore from 'better-queue-memory'
-import axios, { AxiosResponse } from 'axios'
+import axios from 'axios'
+import { Config, Parsers, Handlers, Request, Event } from './types'
 
-interface Request {
-  url: string
-  method: string
-  body?: object
-  authorization?: boolean
-  override?: { headers?: Record<string, unknown> }
-}
-
-export default class Client {
-  private getClientVersion: () => string
-  private getAuthorization: () => string | undefined
-  private isAuthorizationExpired: () => boolean
-  private handleRefresh: () => unknown
-  private handleError: <T>(response: AxiosResponse<T>) => unknown
-  private handleInvalidAuthorization: <T>(response: AxiosResponse<T>) => unknown
+export default class Client extends EventEmitter {
+  private config: Config
+  private parsers: Parsers
+  private handlers: Handlers
   private queue: Queue
 
-  constructor(
-    getClientVersion: () => string,
-    getAuthorization: () => string | undefined,
-    isAuthorizationExpired: () => boolean,
-    handleRefresh: () => {},
-    handleError: <T>(response: AxiosResponse<T>) => unknown,
-    handleInvalidAuthorization: <T>(response: AxiosResponse<T>) => unknown,
-  ) {
-    this.getClientVersion = getClientVersion
-    this.getAuthorization = getAuthorization
-    this.isAuthorizationExpired = isAuthorizationExpired
-    this.handleRefresh = handleRefresh
-    this.handleError = handleError
-    this.handleInvalidAuthorization = handleInvalidAuthorization
+  constructor(config: Config, parsers: Parsers, handlers: Handlers) {
+    super()
+
+    this.config = config
+    this.parsers = parsers
+    this.handlers = handlers
     this.queue = new Queue(
       async (call, cb) => {
         try {
@@ -50,12 +34,12 @@ export default class Client {
   private getHeaders = (authorization = false) => {
     const headers = {
       'Content-Type': 'application/json',
-      'X-Client-Version': this.getClientVersion(),
+      'X-Client-Version': this.config.getClientVersion(),
       Authorization: undefined as string | undefined,
     }
 
-    if (authorization && this.getAuthorization()) {
-      headers.Authorization = this.getAuthorization()
+    if (authorization && this.config.getAuthorization()) {
+      headers.Authorization = this.config.getAuthorization()
     }
     return headers
   }
@@ -81,38 +65,30 @@ export default class Client {
     return options
   }
 
-  public call = <T>(request: Request) => {
-    return new Promise((resolve, reject) => {
-      if (request.authorization && this.isAuthorizationExpired()) {
-        this.queue.push(this.handleRefresh)
+  public call = <T = any>(request: Request) => {
+    return new Promise<T>((resolve, reject) => {
+      if (request.authorization && this.config.getAuthorizationExpired()) {
+        this.queue.push(this.handlers.refresh)
       }
 
       this.queue.push(async () => {
         try {
           const response = await axios<T>(this.getOptions(request))
-          return resolve(this.handleResponse(response))
-        } catch (e) {
+          return resolve(response.data)
+        } catch (e: any) {
+          if (e.response) {
+            if (e.response.status === 401) {
+              this.emit(Event.Unauthorized, e.response)
+            }
+
+            reject(await this.parsers.error(e.response))
+          }
+
           reject(e)
         }
       })
     })
   }
-
-  private handleResponse = <T>(response: AxiosResponse<T>) => {
-    const { status } = response
-
-    switch (status) {
-      case 204:
-        break
-
-      case 200:
-        return response.data
-
-      case 401:
-        return this.handleInvalidAuthorization(response)
-
-      default:
-        return this.handleError(response)
-    }
-  }
 }
+
+export * from './types'