+++ packages/sdk/src/auth/index.test.ts
@@ -83,7 +83,7 @@ describe('AuthSDK', () => {
   it('refreshes and updates token', async () => {
     const sdk = new AuthSDK(config, storage)
 
-    mockTokenResponse({ expires_in: -1 })
+    mockTokenResponse({ expires_in: 0 })
 
     await sdk.makeSession('code')
 
@@ -122,6 +122,12 @@ describe('AuthSDK', () => {
     mockCache()
     mockTokenResponse({ access_token: 'new_access_token', expires_in: 0 })
 
+    const mockSeconds = 1466424490000
+    const mockDateInstance = new Date(mockSeconds)
+    const mockDateConstructor = jest
+      .spyOn(global, 'Date')
+      .mockImplementation(() => mockDateInstance)
+
     const sdk = new AuthSDK(config, storage)
     await sdk.restoreFromCache()
     await sdk.refreshTokens()
@@ -138,9 +144,11 @@ describe('AuthSDK', () => {
         },
         oauthCode: 'code',
         oauthConfig: config,
-        lastUpdateTimestamp: new Date().getTime(),
+        lastUpdateTimestamp: mockSeconds,
       }),
     )
+
+    mockDateConstructor.mockRestore()
   })
 
   it('refreshes when force', async () => {
@@ -155,7 +163,7 @@ describe('AuthSDK', () => {
   })
 
   it('skips refresh if tokens are up-to-date', async () => {
-    mockCache({ lastUpdateTimestamp: new Date().getTime() - 50 })
+    mockCache({ lastUpdateTimestamp: new Date().getTime() - 3600 * 1000 + 1 })
     mockTokenResponse({ access_token: 'new_access_token' })
 
     const sdk = new AuthSDK(config, storage)