+++ packages/sdk/src/auth/utils.ts
@@ -1,12 +1,13 @@
 export const isLastUpdateTimestampExpired = (
   lastUpdateTimestamp: number | undefined,
-  expiration = 0,
+  expirationInSeconds = 0,
 ) => {
   if (!lastUpdateTimestamp) {
     return true
   }
 
   const now = new Date()
+  const expirationInMiliseconds = expirationInSeconds * 1000
 
-  return Math.abs(lastUpdateTimestamp - now.getTime()) > expiration
+  return now.getTime() - lastUpdateTimestamp >= expirationInMiliseconds
 }