+++ packages/api/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @tryrolljs/api
 
+## 0.0.10
+
+### Patch Changes
+
+- [#141](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/141) [`9f6e363`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/9f6e3637330ac931b08fa1d21ab9d05cb18a6893) Thanks [@teimurjan](https://github.com/teimurjan)! - Update engines to match any 16 node
+
+## 0.0.9
+
+### Patch Changes
+
+- [#137](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/137) [`0236b9c`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/0236b9c45160cdebfcbf8151279e6416b398f8d0) Thanks [@teimurjan](https://github.com/teimurjan)! - Update call method typing
+
+## 0.0.8
+
+### Patch Changes
+
+- [#132](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/132) [`1902a60`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/1902a605bc5337ac1a572aa658a86649c2ca963b) Thanks [@teimurjan](https://github.com/teimurjan)! - Refactor client & extend from event emitter
+
+## 0.0.7
+
+### Patch Changes
+
+- [#130](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/130) [`8a5126f`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/8a5126fef714d7bdbc458662689dc31fa46a8cfb) Thanks [@teimurjan](https://github.com/teimurjan)! - Update request token to use post payload instead of query params
+
+## 0.0.6
+
+### Patch Changes
+
+- [#128](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/128) [`1a0b73f`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/1a0b73f4f02adb19688580f9d0633b3023ed5d2c) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix expiration check
+
+## 0.0.5
+
+### Patch Changes
+
+- [#121](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/121) [`8c5844e`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/8c5844ee42c97a5243e67952b7a3fe743ba11570) Thanks [@teimurjan](https://github.com/teimurjan)! - Make API client to accept async handlers
+
+## 0.0.4
+
+### Patch Changes
+
+- [#114](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/114) [`5a81f57`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/5a81f5731e59ef8fe69a141e080f4a7f9717d9f4) Thanks [@teimurjan](https://github.com/teimurjan)! - Update error handling
+
+## 0.0.3
+
+### Patch Changes
+
+- [#109](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/109) [`a3cf603`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/a3cf603cb3fe83ba6b6800c99ac5186c106619b2) Thanks [@teimurjan](https://github.com/teimurjan)! - Add invalid authorization handler
+
 ## 0.0.2
 
 ### Patch Changes