+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tryrolljs/design-system
 
+## 0.1.11
+
+### Patch Changes
+
+- [#146](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/146) [`82a505a`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/82a505ae7850790d624bb1a7d32c43ba215d3438) Thanks [@diegonzs-roll](https://github.com/diegonzs-roll)! - Prevent scroll when modal is open
+
+## 0.1.10
+
+### Patch Changes
+
+- [#141](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/141) [`9f6e363`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/9f6e3637330ac931b08fa1d21ab9d05cb18a6893) Thanks [@teimurjan](https://github.com/teimurjan)! - Update engines to match any 16 node
+
+- [#143](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/143) [`9f8fbb4`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/9f8fbb4bf3eeeb3e6888a2428c9d3b40a758120f) Thanks [@diegonzs-roll](https://github.com/diegonzs-roll)! - Remove action from invalid network banner
+
 ## 0.1.9
 
 ### Patch Changes