+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,73 @@
 # @tryrolljs/sdk
 
+## 0.0.11
+
+### Patch Changes
+
+- [#144](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/144) [`125611e`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/125611ef197e3f2368b90885d362da3c5fc8f5f9) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix refresh condition
+
+## 0.0.10
+
+### Patch Changes
+
+- [#141](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/141) [`9f6e363`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/9f6e3637330ac931b08fa1d21ab9d05cb18a6893) Thanks [@teimurjan](https://github.com/teimurjan)! - Update engines to match any 16 node
+
+- Updated dependencies [[`9f6e363`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/9f6e3637330ac931b08fa1d21ab9d05cb18a6893)]:
+  - @tryrolljs/api@0.0.10
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`0236b9c`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/0236b9c45160cdebfcbf8151279e6416b398f8d0)]:
+  - @tryrolljs/api@0.0.9
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`1902a60`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/1902a605bc5337ac1a572aa658a86649c2ca963b)]:
+  - @tryrolljs/api@0.0.8
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`8a5126f`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/8a5126fef714d7bdbc458662689dc31fa46a8cfb)]:
+  - @tryrolljs/api@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- [#128](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/128) [`1a0b73f`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/1a0b73f4f02adb19688580f9d0633b3023ed5d2c) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix expiration check
+
+- Updated dependencies [[`1a0b73f`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/1a0b73f4f02adb19688580f9d0633b3023ed5d2c)]:
+  - @tryrolljs/api@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`8c5844e`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/8c5844ee42c97a5243e67952b7a3fe743ba11570)]:
+  - @tryrolljs/api@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- [#114](https://github.com/TuringAdvisoryGroup/tryrolljs/pull/114) [`5a81f57`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/5a81f5731e59ef8fe69a141e080f4a7f9717d9f4) Thanks [@teimurjan](https://github.com/teimurjan)! - Update error handling
+
+- Updated dependencies [[`5a81f57`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/5a81f5731e59ef8fe69a141e080f4a7f9717d9f4)]:
+  - @tryrolljs/api@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`a3cf603`](https://github.com/TuringAdvisoryGroup/tryrolljs/commit/a3cf603cb3fe83ba6b6800c99ac5186c106619b2)]:
+  - @tryrolljs/api@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes